Minor Improvements

+++ Mortgage.Api/Application/Dtos/ScheduledPaymentDto.cs
@@ -3,8 +3,8 @@ public class ScheduledPaymentDto
     public Guid Id {get; set;}
     public int Numer_Raty {get; set;}
     public string? Data_Płatności {get; set;}
-    public double Wysokość_Raty {get; set;}
-    public double Kwota_Odsetek {get; set;}
-    public double Kwota_Kapitału {get; set;}
-    public double Pozostało_Do_Spłaty {get; set;}
+    public decimal Wysokość_Raty {get; set;}
+    public decimal Kwota_Odsetek {get; set;}
+    public decimal Kwota_Kapitału {get; set;}
+    public decimal Pozostało_Do_Spłaty {get; set;}
 }
